Big reorganization of the attribute types.
The concrete types no longer are prefixed with '_'.
The type union reflecting the accepted argument types
are now suffixed with 'Spec' (for 'specification').

The concrete types which. are stored in different formats
reveal that via a .size property.

Subtracting points now gives the vector to the second point.
To get distance, use the .length property, or the
Point.distance(p1, p2) static function.